refactor: decompose ModelManagementSheet into subviews
